+++ ToyBox/Classes/Features/SettingsTab/SettingsFeatureTab.cs
@@ -69,6 +69,7 @@ public partial class SettingsFeaturesTab : FeatureTab {
         AddFeature(new CharacterPickerNearbyRangeSetting(), m_OtherText);
         AddFeature(new LogHotkeysToCombatLogSetting(), m_OtherText);
         AddFeature(new ImguiColorFixFeature(), m_OtherText);
+        AddFeature(new GlyphSettingsFeature(), m_OtherText);
 
         AddFeature(new LanguagePickerFeature(), m_LanguageText);
     }