+++ ToyBox/Classes/Infrastructure/Settings/GeneralSettings.cs
@@ -46,6 +46,8 @@ public class GeneralSettings : AbstractJsonSettings {
     public bool EnableLogHotkeysToCombatLog = false;
     public bool EnableShowRiskyToggles = false;
     public bool EnableImguiColorFix = true;
+    public bool CheckForGlyphSupport = true;
+    public bool UseDefaultGlyphs = true;
 
     // - Browser
     public int PageLimit = 25;